move `initialize()` function to `elements/index.js`
