+++ .github/workflows/prettier-standard.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@master
-    - name: Setup node
+    - name: Setup Node
       uses: actions/setup-node@v2
       with:
         node-version: '12'
