+++ .github/workflows/tests.yml
@@ -33,12 +33,12 @@ jobs:
     name: Node v${{ matrix.node }}
     steps:
       - uses: actions/checkout@v2
-      - name: Setup node
+      - name: Setup Node
         uses: actions/setup-node@v2
         with:
           node-version: ${{ matrix.node }}
           cache: 'yarn'
-      - name: Build javascript package
+      - name: Build JavaScript package
         run: |
           yarn install
           yarn build
